scraper/requirements.txt
# Scraping
playwright>=1.44.0
playwright-stealth>=1.0.6

# Parsing HTML
beautifulsoup4>=4.12.0
lxml>=5.2.0

# HTTP async
httpx>=0.27.0

# DNS / SMTP verification
dnspython>=2.6.0

# Data
pandas>=2.2.0
requests>=2.31.0
